Added ability to upload a photo with Memory
Created a String var in Memory class called imageUri and this is the String representation of the uri for the image uploaded into the firebase Storage.  I made the image folder have the UID of the authorized user so that it is easy to know who's photos belong to whom for deleting purposes.

Now I just need to get a page ready for displaying data so I can see how it displays and then incorporate Praneet's code for recycler view