ImageView isn't changing - rest of code is working
At this point the Image Uri is uploading correctly.  The data is all saving correctly and pulling correctly to and from firebase.  However, when I go tot my EditMemoryActivity screen, and try to set the ImageUri with the value from Firebase, it makes the image blank.  If I hard code an image in, then it shows up.  Once I try to change it with the Uri it is blank.

I don't yet have the code working to allow me to edit or delete the item